# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PROJECT_NAME API",
    "description": "PROJECT_NAME 对外 REST API（以 OpenAPI 作为规范中心）。",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/security/password": {
      "patch": {
        "tags": [
          "security"
        ],
        "operationId": "patch_current_user_password_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchCurrentUserPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "修改当前用户密码成功（无 body）"
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "未登录或 Token 无效 / 当前密码错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "当前用户不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponseBody": {
        "type": "object",
        "description": "失败时的统一错误体（与 `AppError` 的序列化保持一致）。",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
...
      "PatchCurrentUserPasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password",
            "maxLength": 256,
            "minLength": 1
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "maxLength": 256,
            "minLength": 8
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}